scraper/requirements.txt
requests
beautifulsoup4
pyarrow
pyyaml
